Avance controlador de ordenDespacho y implementacion de servicios

+++ src/main/java/pe/com/pathOrder/demoxd/repository/FacturaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pe.com.pathOrder.demoxd.model.Factura;
 
-public interface FacturaRepository extends JpaRepository<Factura, Long>{
+public interface FacturaRepository extends JpaRepository<Factura, Integer>{
 
 }
